[Mac] A checkmark is briefly visible next to the webpage context menu items when they're clicked https://bugs.webkit.org/show_bug.cgi?id=146284 <rdar://problem/20297182>
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2015-06-24
Reviewed by Darin Adler.

The default value of [NSPopUpButtonCell altersStateOfSelectedItem] is YES.
This makes the popup menu items checkable by default. We need to change the
value of this attribute to NO when creating the popup menu, so the items
are uncheckable. The value of this property needs to change back to YES
if context menu is about selecting an option and not executing a command.

* UIProcess/mac/WebContextMenuProxyMac.mm:
(WebKit::WebContextMenuProxyMac::populate):

Identifier: 164361@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185922 268f45cc-cd09-0410-ab3c-d52691b4dbfc