Add a new color -apple-wireless-playback-target-active https://bugs.webkit.org/show_bug.cgi?id=146269 <rdar://problem/21465189>
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2015-06-24
Reviewed by Dean Jackson.

Source/WebCore:

At the moment we use -apple-system-blue to indicate an Airplay route is
active. This might not always be blue, so we need to add another system
color with a name that will allow us to change the underlying color.
Initially -apple-wireless-playback-target-active will be an alias to
-apple-system-blue.

Existing tests are modified to include the new color value.

* Modules/mediacontrols/mediaControlsApple.js:
(Controller.prototype.updateWirelessTargetPickerButton):
* Modules/mediacontrols/mediaControlsiOS.css:
(video::-webkit-media-controls-wireless-playback-picker-button.playing):
Use the new color value in the media controls.

* css/CSSParser.cpp:
(WebCore::isAppleLegacyCssValueKeyword): Returns true if the value keyword
has an '-apple' prefix which has to be replaced by '-webkit'

(WebCore::cssValueKeywordID): Call a new function which handles all the
'-apple' legacy keywords.

* css/CSSValueKeywords.in: Define the new color keyword.

* rendering/RenderThemeIOS.mm:
(WebCore::RenderThemeIOS::systemColor):
* rendering/RenderThemeMac.mm:
(WebCore::RenderThemeMac::systemColor):
Define the value of the new color keyword to be an alias of systemBlueColor.

LayoutTests:

Modify existing tests to include the new color value.

* fast/css/apple-system-colors-expected.txt:
* fast/css/apple-system-colors.html:
* platform/ios-simulator/fast/css/apple-system-colors-expected.txt:
* platform/mac-mavericks/fast/css/apple-system-colors-expected.txt:
* platform/mac/fast/css/apple-system-colors-expected.txt:

Identifier: 164373@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185934 268f45cc-cd09-0410-ab3c-d52691b4dbfc