Viewport units are wrong when scaled in 2-up mode, cause content to hop around on apple.com/music https://bugs.webkit.org/show_bug.cgi?id=146322 <rdar://problem/21413884>
Reviewed by Simon Fraser.

* page/FrameView.cpp:
(WebCore::FrameView::viewportSizeForCSSViewportUnits):
Use the fixed layout size, if enabled, instead of the visibleContentRect,
for the viewport unit size. This ensures that viewport units take up the whole
fixed-layout viewport, not just the size of the view scaled by the page scale
(which visibleContentRect provides).

* WebProcess/WebPage/mac/TiledCoreAnimationDrawingArea.mm:
(WebKit::TiledCoreAnimationDrawingArea::scaleViewToFitDocumentIfNeeded):
Inflate the fixed layout height (by the viewScale) so that it's the size of the view.

* fast/fixed-layout/fixed-layout-expected.txt:
Rebaseline now that viewport units behave correctly.


Identifier: 164403@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185966 268f45cc-cd09-0410-ab3c-d52691b4dbfc