WKWebView snapshots have the wrong scale after rotation https://bugs.webkit.org/show_bug.cgi?id=146476 <rdar://problem/18345247>
Reviewed by Anders Carlsson.

* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _didCommitLayerTree:]):
"Coordinate" -> "Coordinates".

(-[WKWebView _snapshotRect:intoImageOfWidth:completionHandler:]):
Defer any snapshotting that happens during a resize until after
the resize completes. This will ensure that (in the case of an IOSurface
snapshot) the tiles are up to date, and (in the case of a software snapshot)
that our understanding of the scale and scroll offset of the WKContentView
are up to date, so that we can correctly convert from view to content coordinates
in the UI process.

(-[WKWebView _endAnimatedResize]):
Perform the deferred snapshotting after the resize completes.


Identifier: 164608@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@186194 268f45cc-cd09-0410-ab3c-d52691b4dbfc