bmalloc: Shrink the super chunk size https://bugs.webkit.org/show_bug.cgi?id=146519
Reviewed by Andreas Kling.

We have lots of reports of crashing due to failed VM allocation on iOS.
(This VM limit on iOS is usually 1GB-2GB, and has been as low as 256MB.)

Shrink the super chunk size in case fragmentation is the reason for
VM allocation failure.

This has the downside that >= 2MB allocations will now be super slow,
but they are also super rare (as in never on most websites), so this
is probably an OK tradeoff.

* bmalloc/Sizes.h:


Identifier: 164654@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@186242 268f45cc-cd09-0410-ab3c-d52691b4dbfc