Cleanup: Remove default constructor for WebCore::Pair() https://bugs.webkit.org/show_bug.cgi?id=146856 <rdar://problem/21773212>
Reviewed by Darin Adler.

Remove default constructor for WebCore::Pair() and the setters Pair::set{First, Second}()
since they are not used and it seems weird to keep the latter after the removal of the former.
Should it turn out that we find a use for having a default constructor and/or setters then
we can revert this change.

* css/Pair.h:
(WebCore::Pair::create): Deleted.
(WebCore::Pair::setFirst): Deleted.
(WebCore::Pair::setSecond): Deleted.
(WebCore::Pair::Pair): Deleted.


Identifier: 164941@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@186704 268f45cc-cd09-0410-ab3c-d52691b4dbfc